MovieDetails: Audio: Look for default audio track
Currently, we always assume the first audio track on the list is
the default audio track for the movie. This does not support all
configurations of metadata smoothly.

As such, this patch updates the logic to first look for a default
audio track (the first on the list if there are multiple) and then,
if none is found, revert back to the first available audio track.